move shapes into separate classes/files to make it easier to add and remove them
