+++ package.json
@@ -9,6 +9,7 @@
     "react-bootstrap": "^0.31.0",
     "react-dom": "^15.6.1",
     "react-redux": "^5.0.5",
+    "reduce-reducers": "^1.0.4",
     "react-scripts": "2.1.5",
     "redux": "^3.7.1",
     "reselect": "^4.0.0",
