+++ src/shapes/Vicious1Vertices.js
